// src/common/enums/role.enum.ts
export enum RoleEnum {
  USER = 'user',
  ADMIN = 'admin',
  MERCHANT = 'merchant',
  ADMIN_SUPERADMIN = 'admin_superadmin',
  ADMIN_CLIENT_READ_ONLY = 'admin_client_read_only',
  ADMIN_CLIENT_MANAGER = 'admin_client_manager',
  ADMIN_CONFIG_READ_ONLY = 'admin_config_read_only',
  ADMIN_CONFIG_MANAGER = 'admin_config_manager',
}